Updated timeout propagation test to actually test that it's getting propagated
